Split exceptions into submodule
