+++ praw/exceptions/base.py
@@ -1,0 +1,5 @@
+"""Container for the base exception."""
+
+
+class PRAWException(Exception):
+    """The base PRAW Exception that all other exception classes extend."""
